Refactor spelling/grammar workflows to use shared utilities
- Extract duplicate functions (splitIntoChunks, getCharacterOffsetForLine, getErrorGroupEmoji, getErrorTypeLabel) to utils.ts
- Import and use shared functions in both workflow files
- Add proper type imports to replace 'any' types
- Clean up type safety with ProcessedResults and ErrorGroup imports
- Reduce code duplication between parallel and sequential workflows
